refactor: move openapi tests to a dedicated file
